requirements.txt
discord.py==2.3.2
Mastodon.py==1.8.1
pluralkit==1.1.5
python-dateutil==2.8.2
python-magic==0.4.24
pytz==2022.1
PyYAML==6.0.1
urllib3==1.26.5

